backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-vertexai>=1.0.0
langchain-community>=0.2.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
